updates for TX16mk3 800x480 resolution
