+++ sdcard/c800x480/WIDGETS/BattAnalog/lib_log.lua
@@ -2,6 +2,7 @@ local app_name, script_dir = ...
 
 local ENABLE_LOG_TO_CONSOLE = true
 local ENABLE_LOG_TO_FILE    = false
+local ENABLE_LOG_TO_SERIAL  = false
 
 
 local M = {}
@@ -17,6 +18,7 @@ local log = {
     outfile = script_dir .. "/app.log",
     enable_file = ENABLE_LOG_TO_FILE,
     enable_console = ENABLE_LOG_TO_CONSOLE and is_simulator(),
+    enable_serial_dbg = ENABLE_LOG_TO_SERIAL,
     current_level = nil,
 
     -- func
@@ -61,7 +63,7 @@ local function tostring(...)
 end
 
 function M.do_log(iLevel, ulevel, fmt, ...)
-    if log.enable_console == false and log.enable_file == false then
+    if log.enable_console == false and log.enable_file == false and log.enable_serial_dbg == false then
         return
     end
 
@@ -91,6 +93,12 @@ function M.do_log(iLevel, ulevel, fmt, ...)
         io.write(fp, msg2 .. "\n")
         io.close(fp)
     end
+
+    -- Output to log file
+    if log.enable_serial_dbg == true then
+        serialWrite(msg2.."\r\n") -- 115200 bps
+    end
+
 end
 
 function M.trace(fmt, ...)
